First part of the refactoring
