+++ src/platform/test/functional/apps/discover/observability/index.ts
@@ -31,9 +31,7 @@ export default function ({ getService, getPageObjects, loadTestFile }: FtrProvid
       });
     });
 
-    loadTestFile(require.resolve('./embeddable/_saved_search_embeddable'));
     loadTestFile(require.resolve('./logs/_get_pagination_config'));
-    loadTestFile(require.resolve('./embeddable/_get_doc_viewer'));
     loadTestFile(require.resolve('./logs/_get_doc_viewer'));
     loadTestFile(require.resolve('./logs/_get_recommended_fields'));
     loadTestFile(require.resolve('./logs/_summary_column'));
