+++ src/platform/test/functional/apps/discover/observability/logs/_get_doc_viewer.ts
@@ -14,62 +14,21 @@ import type { FtrProviderContext } from '../../ftr_provider_context';
 import { MORE_THAN_1024_CHARS, STACKTRACE_MESSAGE } from '../const';
 
 export default function ({ getService, getPageObjects }: FtrProviderContext) {
-  const PageObjects = getPageObjects(['common', 'discover']);
+  const PageObjects = getPageObjects(['common', 'discover', 'dashboard', 'header']);
   const testSubjects = getService('testSubjects');
   const dataGrid = getService('dataGrid');
-  const dataViews = getService('dataViews');
+  const dashboardAddPanel = getService('dashboardAddPanel');
   const synthtrace = getService('synthtrace');
   const queryBar = getService('queryBar');
+  const kibanaServer = getService('kibanaServer');
+  const dataViews = getService('dataViews');
 
   const start = moment().subtract(30, 'minutes').valueOf();
   const end = moment().add(30, 'minutes').valueOf();
 
-  describe('extension getDocViewer ', () => {
-    let synthEsLogsClient: LogsSynthtraceEsClient;
-    before(async () => {
-      const { logsEsClient } = synthtrace.getClients(['logsEsClient']);
+  const searchQuery = 'error.stack_trace : * and _ignored : *';
 
-      synthEsLogsClient = logsEsClient;
-
-      await synthEsLogsClient.index([
-        timerange(start, end)
-          .interval('1m')
-          .rate(5)
-          .generator((timestamp: number, index: number) =>
-            log
-              .create()
-              .message('This is a log message')
-              .timestamp(timestamp)
-              .dataset('synth.discover')
-              .namespace('default')
-              .logLevel(index % 2 === 0 ? MORE_THAN_1024_CHARS : 'This is a log message')
-              .defaults({
-                'service.name': 'synth-discover',
-                ...(index % 2 === 0 && { 'error.stack_trace': STACKTRACE_MESSAGE }),
-              })
-          ),
-      ]);
-
-      await PageObjects.common.navigateToActualUrl('discover', undefined, {
-        ensureCurrentUrl: false,
-      });
-
-      // Required as some other test switches data view to metric-*
-      await dataViews.switchTo('All logs');
-
-      await queryBar.setQuery('error.stack_trace : * and _ignored : *');
-      await queryBar.submitQuery();
-      await PageObjects.discover.waitUntilTabIsLoaded();
-    });
-
-    after(async () => {
-      await synthEsLogsClient.clean();
-    });
-
-    afterEach(async () => {
-      await dataGrid.closeFlyout();
-    });
-
+  function defineDocViewerTests(rowIndices: { qualityIssue: number; stacktrace: number }) {
     describe('renders docViewer', () => {
       it('should open the flyout with stacktrace and quality issues accordion closed when expand is clicked', async () => {
         await dataGrid.clickRowToggle({ rowIndex: 0 });
@@ -80,7 +39,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Quality Issues accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewDegradedFieldsAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
@@ -88,7 +46,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Stacktrace accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewStacktraceAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
@@ -103,7 +60,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Quality Issues accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewDegradedFieldsAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
@@ -111,7 +67,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Stacktrace accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewStacktraceAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
@@ -126,7 +81,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Quality Issues accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewDegradedFieldsAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
@@ -134,7 +88,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
 
         // Stacktrace accordion to be present and collapsed
         await testSubjects.existOrFail('unifiedDocViewLogsOverviewStacktraceAccordion');
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
@@ -144,25 +97,24 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
       it('should keep old accordion open when 1st stacktrace and then quality issue control for the same row is clicked', async () => {
         await dataGrid.clickStacktraceLeadingControl(0);
 
+        // 1st stack trace accordion should be opened and quality issues should be closed
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
         );
 
         // Clicking on Quality Issue control of the same row while the Flyout is still open
-
         await dataGrid.clickQualityIssueLeadingControl(0);
 
+        // Expect the previous one to stay open and new one to also open. This shows component did not remount
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
@@ -172,25 +124,24 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
       it('should toggle to quality issue accordion when 1st stacktrace and then quality issue control is clicked for different row', async () => {
         await dataGrid.clickStacktraceLeadingControl(0);
 
+        // 1st stack trace accordion should be opened and quality issues should be closed
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
         );
 
-        // Clicking on Quality Issue control of the same row while the Flyout is still open
-
+        // Clicking on Quality Issue control of a different row while the Flyout is still open
         await dataGrid.clickQualityIssueLeadingControl(1);
 
+        // Expect toggle to have happened
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
@@ -200,11 +151,11 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
       it('should keep old accordion open when 1st quality issue and then stacktrace control for the same row is clicked', async () => {
         await dataGrid.clickQualityIssueLeadingControl(0);
 
+        // 1st quality issues accordion should be opened and stacktrace should be closed
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
@@ -213,11 +164,11 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
         // Clicking on Stacktrace control of the same row while the Flyout is still open
         await dataGrid.clickStacktraceLeadingControl(0);
 
+        // Expect the previous one to stay open and new one to also open. This shows component did not remount
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
@@ -225,26 +176,26 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
       });
 
       it('should toggle to stacktrace accordion when 1st quality issue and then stacktrace control is clicked for different row', async () => {
-        await dataGrid.clickQualityIssueLeadingControl(0);
+        await dataGrid.clickQualityIssueLeadingControl(rowIndices.qualityIssue);
 
+        // 1st quality issues accordion should be opened and stacktrace should be closed
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
         );
 
-        // Clicking on Stacktrace control of the same row while the Flyout is still open
-        await dataGrid.clickStacktraceLeadingControl(1);
+        // Clicking on Stacktrace control of a different row while the Flyout is still open
+        await dataGrid.clickStacktraceLeadingControl(rowIndices.stacktrace);
 
+        // Expect toggle to have happened
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
@@ -265,7 +216,6 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
@@ -279,14 +229,13 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
         const jsonTabButton = await testSubjects.find('docViewerTab-doc_view_source');
         await jsonTabButton.click();
 
-        // Click to open Quality Issue control on the same row
+        // Click to open Quality Issue control on a different row
         await dataGrid.clickQualityIssueLeadingControl(1);
 
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'true'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'false'
@@ -300,14 +249,13 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
         const jsonTabButton = await testSubjects.find('docViewerTab-doc_view_source');
         await jsonTabButton.click();
 
-        // Click to open Quality Issue control on the same row
+        // Click to open Stacktrace control on the same row
         await dataGrid.clickStacktraceLeadingControl(0);
 
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
@@ -321,19 +269,113 @@ export default function ({ getService, getPageObjects }: FtrProviderContext) {
         const jsonTabButton = await testSubjects.find('docViewerTab-doc_view_source');
         await jsonTabButton.click();
 
-        // Click to open Quality Issue control on the same row
+        // Click to open Stacktrace control on a different row
         await dataGrid.clickStacktraceLeadingControl(1);
 
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewDegradedFieldsAccordion',
           'false'
         );
-
         await testSubjects.waitForAccordionState(
           'unifiedDocViewLogsOverviewStacktraceAccordion',
           'true'
         );
       });
+    });
+  }
+
+  describe('getDocViewer', () => {
+    let synthEsLogsClient: LogsSynthtraceEsClient;
+
+    before(async () => {
+      const { logsEsClient } = synthtrace.getClients(['logsEsClient']);
+      synthEsLogsClient = logsEsClient;
+
+      await synthEsLogsClient.index([
+        timerange(start, end)
+          .interval('1m')
+          .rate(5)
+          .generator((timestamp: number, index: number) =>
+            log
+              .create()
+              .message('This is a log message')
+              .timestamp(timestamp)
+              .dataset('synth.discover')
+              .namespace('default')
+              .logLevel(index % 2 === 0 ? MORE_THAN_1024_CHARS : 'This is a log message')
+              .defaults({
+                'service.name': 'synth-discover',
+                ...(index % 2 === 0 && { 'error.stack_trace': STACKTRACE_MESSAGE }),
+              })
+          ),
+      ]);
+    });
+
+    after(async () => {
+      await synthEsLogsClient.clean();
+    });
+
+    describe('in Discover', () => {
+      before(async () => {
+        await PageObjects.common.navigateToActualUrl('discover', undefined, {
+          ensureCurrentUrl: false,
+        });
+
+        // Required as some other test switches data view to metric-*
+        await dataViews.switchTo('All logs');
+
+        await queryBar.setQuery(searchQuery);
+        await queryBar.submitQuery();
+        await PageObjects.discover.waitUntilTabIsLoaded();
+      });
+
+      afterEach(async () => {
+        await dataGrid.closeFlyout();
+      });
+
+      defineDocViewerTests({ qualityIssue: 0, stacktrace: 1 });
+    });
+
+    describe('in Dashboard saved search embeddable', () => {
+      before(async () => {
+        await PageObjects.common.navigateToActualUrl('discover', undefined, {
+          ensureCurrentUrl: false,
+        });
+
+        // Required as some other test switches data view to metric-*
+        await dataViews.switchTo('All logs');
+
+        await queryBar.setQuery(searchQuery);
+        await queryBar.submitQuery();
+        await PageObjects.discover.waitUntilSearchingHasFinished();
+
+        // Required to access Dashboard page
+        await dataViews.createFromSearchBar({
+          name: 'logs-synth',
+        });
+
+        await PageObjects.discover.saveSearch('synth-search-doc-viewer');
+
+        await PageObjects.dashboard.navigateToApp();
+        await PageObjects.dashboard.gotoDashboardLandingPage();
+        await PageObjects.dashboard.clickNewDashboard();
+
+        await dashboardAddPanel.addSavedSearch('synth-search-doc-viewer');
+        await PageObjects.header.waitUntilLoadingHasFinished();
+        await PageObjects.dashboard.waitForRenderComplete();
+      });
+
+      after(async () => {
+        await kibanaServer.savedObjects.clean({
+          types: ['search', 'index-pattern', 'dashboard'],
+        });
+      });
+
+      afterEach(async () => {
+        await dataGrid.closeFlyout();
+      });
+
+      defineDocViewerTests({ qualityIssue: 1, stacktrace: 2 });
     });
   });
 }
